// finance_client/src/app/user/data/types/customer-types/customer.ts
export interface Customer {
  buyerAddress?: string;
  buyerBusinessName?: string;
  buyerCitizenship?: string;
  buyerEmail?: string;
  buyerLegalName?: string;
  buyerLinePhone?: string;
  buyerMobilePhone?: string;
  buyerNinBrn?: string;
  buyerPassportNum?: string;
  buyerPlaceOfBusi?: string;
  buyerReferenceNo?: string;
  buyerSector?: string;
  buyerTin?: string;
  buyerType?: string;
  cellVillage?: string;
  customerAccount?: string;
  deletedBy?: string;
  deletedFlag?: string;
  deletedTime?: Date;
  district?: string;
  divisionSubcounty?: string;
  effectiveRegistrationDate?: string;
  id?: number;
  meterStatus?: string;
  modifiedBy?: string;
  modifiedFlag?: string;
  modifiedTime?: Date;
  municipalityCounty?: string;
  postedBy?: string;
  postedFlag?: string;
  postedTime?: Date;
  propertyType?: string;
  reason?: string;
  status?: string;
  town?: string;
  verifiedBy?: string;
  verifiedFlag?: string;
  verifiedTime?: Date;
}